* New controller OnlinePayments for information and documents * Registration rules extracted to separate include and special view home/terms * Payment info layout refactored
