+++ Joinrpg/Controllers/HomeController.cs
@@ -20,13 +20,11 @@ namespace JoinRpg.Web.Controllers
         public async Task<ActionResult> Index() =>
             View(await ProjectListManager.LoadModel(false, ProjectsOnHomePage));
 
-        public ActionResult Legal() => View();
+        public ActionResult Terms() => View();
 
         public ActionResult About() => View();
 
         public ActionResult Support() => View();
-
-        public ActionResult Tariffs() => View();
 
         public ActionResult Funding2016() => View();
