+++ Joinrpg/Views/Account/_RegistrationRules.cshtml
@@ -1,19 +1,56 @@
-﻿@model dynamic
-<div class="alert alert-info">
-<ol>
-    <li>Мы рады видеть вас на нашем сайте! Надеемся, что ваш опыт на джойне как игрока, и как мастера будет приятным. Ведите себя прилично и вежливо, и все будет хорошо.</li>
+﻿<ol>
     <li>
-        Если у вас будут какие-то проблемы, пишите в поддержку (<span class="glyphicon glyphicon-envelope" aria-hidden="true"></span> <a href="mailto:support@joinrpg.ru">support@joinrpg.ru</a>
-        /
-        <a href="https://vk.com/joinrpg">
-            <img src="~/Content/Images/vk_icon.png"/>
-        </a>). Мы всегда рады помочь, но обратите внимание: вы ничего не платите нам, и, следовательно, не являетесь нашим клиентом. Мы поддерживаем сайт бесплатно и в свободное время, у нас есть работа, семья, личная жизнь, другие хобби.
+        Мы рады видеть вас на нашем сайте! Надеемся, что ваш опыт на JoinRpg и как игрока (участника),
+        и как мастера (организатора) будет приятным. Ведите себя прилично и вежливо, и все будет хорошо.
     </li>
-    <li>Большую часть игр организуем не мы :-) По всем игровым вопросам обращайтесь к мастерам (организаторам) конкретной игры. Мы не обладаем временем или компетенцией, чтобы понять, будет ли игра хорошей или плохой — пообщайтесь с мастерами и составьте свое впечатление.</li>
-    <li>Мы не сможем вам помочь, если вам не понравится игра или общение с мастерами. Если что — просто выберите себе другую игру вместо этой.</li>
-    <li>Мы не принимаем у игроков взносы, они переводятся напрямую мастерам, мы их никак не трогаем.</li>
-    <li>Мы будем вынуждены заблокировать тех пользователей, которые нарушат общепринятые правила интернетов. Это включает (но не ограничивается): спам, взлом нашего или других сайтов, любые другие действия, которые ставят нас под угрозу технически или юридически. Мы не ограничиваем себя определенным списком или правилами, но постараемся применять блокировки только как крайнюю меру. Пока что такой необходимости не возникало.</li>
-    <li>Если кто-то ведет себя как м...к по отношению к нам, мы не даем таким людям согласия на использование нашего сайта.</li>
+    <li>
+        Если у вас будут какие-то проблемы, пишите в поддержку (на почту <a href="mailto:@Contacts.SupportEmail">@Contacts.SupportEmail</a>
+        или в <a href="@Contacts.VkGroup">группу Вконтакте</a>). Мы всегда рады помочь,
+        но обратите внимание: сервис JoinRpg предоставляется вам бесплатно, «как есть».
+        Мы не гаранитируем вам оперативный ответ 24/7 по любой вашей проблеме и не несем ответственности,
+        если вы решите, что JoinRpg вас чем-то не устраивает.
+    </li>
+    <li>
+        Большую часть игр и других мероприятий организуем не мы :-) По всем игровым вопросам обращайтесь
+        к мастерам (организаторам) конкретного мероприятия. Мы не обладаем временем или компетенцией,
+        чтобы понять, будет ли игра или мероприятие хорошим или плохим — пообщайтесь с мастерами
+        (организаторами) и составьте свое впечатление.
+    </li>
+    <li>
+        Мы не сможем вам помочь, если вам чем-то не понравится игра (мероприятие) или общение
+        с мастерами (организаторами). Если что — просто выберите себе другую игру вместо этой.
+        Мы не несем никакой ответственности за деятельность мастеров (организаторов) и за
+        соответствие игры (мероприятия) вашим ожиданиям.
+    </li>
+    <li>
+        JoinRpg не имеет доступа к информации о ваших банковских картах.
+        <ul>
+            <li>
+                Если мастера (организаторы) включили прием взносов (продажу билетов) через JoinRpg, то денежные
+                средства принимаются через платежный сервис «ПСКБ Онлайн», работающий согласно требованиям стандарта
+                безопасности PCI&nbsp;DSS&nbsp;3.0, все данные платежа передаются через соединение,
+                зашифрованное по технологии SSL. Безопасность гарантирует <a target="_blank" href="https://pscb.ru">АО&nbsp;Банк&nbsp;«ПСКБ»</a>.
+                После поступления денежных средств на расчетный счет ООО «Джойн Регистрация» они
+                передаются в распоряжение мастеров (организаторов).
+            </li>
+            <li>
+                Во всех остальных случаях через JoinRpg взносы не принимаются. Их принимают непосредственно
+                мастера (организаторы).
+            </li>
+        </ul>
+        Если вы не уверены, можно ли сдать взнос через JoinRpg, напишите на <a href="mailto:@Contacts.PaymentsEmail">@Contacts.PaymentsEmail</a>
+        с указанием, о каком мероприятии идет речь.
+    </li>
+    <li>
+        Мы будем вынуждены заблокировать тех пользователей, которые нарушат общепринятые правила Интернета.
+        Это включает (но не ограничивается): спам, взлом JoinRpg или других сайтов, любые другие действия,
+        которые ставят JoinRpg под угрозу технически или юридически. Мы не ограничиваем себя определенным
+        списком или правилами, но постараемся применять блокировки только как крайнюю меру. Пока что такой
+        необходимости не возникало.
+    </li>
+    <li>
+        Если кто-то ведет себя как м...к по отношению к нам, то мы не даем таким людям согласия на
+        использование нашего сайта.
+    </li>
     <li>Рады вас видеть и удачных вам игр!</li>
 </ol>
-</div>
