+++ Joinrpg/Views/Finances/_PaymentLegalInfoPartial.cshtml
@@ -1,21 +1,16 @@
-﻿<div class="join alert alert-info">
-    <i class="glyphicon glyphicon-info-sign"></i>
-    <div style="font-size: smaller;">
-        <p>
-            После нажатия кнопки «Оплатить» вы будете перенаправлены на страницу платежного сервиса
-            «ПСКБ Онлайн» для ввода данных своей банковской карты (поддерживаются карты Master Card, Visa или МИР).
-            Если банк, выдавший вам карту, требует дополнительную проверку, то произойдет переадресация на страницу
-            банка для ввода одноразового пароля (технология 3D Secure).
-        </p>
-        <p>
-            JoinRpg.Ru не имеет доступа к данным вашей карты, и соответственно не обрабатывает и не хранит их,
-            они обрабатываются платежным сервисом «ПСКБ Онлайн» согласно требованиям стандарта безопасности PCI&nbsp;DSS&nbsp;3.0,
-            передаются через соединение, зашифрованное по технологии SSL. Безопасность гарантирует
-            <a target="_blank" href="https://pscb.ru">АО&nbsp;Банк&nbsp;«ПСКБ»</a>.
-        </p>
-        <p>
-            Успешная оплата будет автоматически отмечена в базе, а вы получите чек на свою электронную почту.
-            Дополнительно информировать организаторов (мастеров) или отмечать оплату вручную не требуется.
-        </p>
-    </div>
-</div>
+﻿<p>
+    После нажатия кнопки «Оплатить» вы будете перенаправлены на страницу платежного сервиса
+    «ПСКБ Онлайн» для ввода данных своей банковской карты (поддерживаются карты Master Card, Visa или МИР).
+    Если банк, выдавший вам карту, требует дополнительную проверку, то произойдет переадресация на страницу
+    банка для ввода одноразового пароля (технология 3D Secure).
+</p>
+<p>
+    JoinRpg.Ru не имеет доступа к данным вашей карты, и соответственно не обрабатывает и не хранит их,
+    они обрабатываются платежным сервисом «ПСКБ Онлайн» согласно требованиям стандарта безопасности PCI&nbsp;DSS&nbsp;3.0,
+    передаются через соединение, зашифрованное по технологии SSL. Безопасность гарантирует
+    <a target="_blank" href="https://pscb.ru">АО&nbsp;Банк&nbsp;«ПСКБ»</a>.
+</p>
+<p>
+    Успешная оплата будет автоматически отмечена в базе, а вы получите чек на свою электронную почту.
+    Дополнительно информировать организаторов (мастеров) или отмечать оплату вручную не требуется.
+</p>
