--- Joinrpg/Views/Home/Tariffs.cshtml
@@ -1,30 +1,0 @@
-﻿<h1>Тарифы</h1>
-
-<div class="panel panel-default">
-    <div class="panel-heading">Тариф «Бесплатный»</div>
-    <div class="panel-body">
-        <p>Вы можете пользоваться сервисом JoinRpg бесплатно.</p>
-        <p>Вы осуществляете сбор взносов (продажу билетов) любым удобным вам способом,
-            самостоятельно ведете учет своих финансов, в том числе самостоятельно отмечаете факт
-            сдачи взноса вручную.</p>
-    </div>
-</div>
-
-<div class="panel panel-default">
-    <div class="panel-heading">Тариф «Простой»</div>
-    <div class="panel-body">
-        <p>Организатором должно выступать юридическое лицо либо индивидуальный предприниматель.
-            С организатором заключается агентский договор.</p>
-        <p>Вознаграждение ООО «Джойн Регистрация» составляет <b>5%</b> с каждого принятого взноса (проданного билета),
-            плюс одноразовый сервисный сбор в размере <b>5000</b> рублей.</p>
-        <p>Важно: организатор вправе собирать взносы (продавать билеты) иными способами, однако стоимости
-            эквивалентных по типу взносов (билетов) должны быть одинаковыми вне зависимости от того,
-            сдается взнос (покупается билет) через JoinRpg, или иными способами.</p>
-        <p>Каждый сданный взнос (проданный билет) автоматически отмечается в регистрационной заявке.
-            Ручная обработка не требуется.</p>
-        <p>Перечисление средств на расчетный счет организатора производится по запросу не чаще 1 раза в месяц.
-            За 2 месяца до мероприятия не чаще 1 раза в 2 недели, за 1 месяц до мероприятия не чаще 1 раза в неделю.</p>
-        <p>Организаторам на этом тарифе оказывается приоритетная поддержка.</p>
-        <p style="margin-top: 1em;">Хотите подключить сбор взносов (продажу билетов) прямо из заявки в JoinRpg? Срочно <a href="mailto:@Contacts.PaymentsEmail">пишите нам</a>!</p>
-    </div>
-</div>
