+++ Joinrpg/Views/Home/Terms.cshtml
@@ -1,0 +1,3 @@
+﻿<h1>Правила использования сайта</h1>
+
+@Html.Partial("..\\Account\\_RegistrationRules")
