+++ Joinrpg/Views/OnlinePayments/Index.cshtml
@@ -82,7 +82,7 @@
 <h2>Документы</h2>
 
 <div>
-    <h3><a href="@Documents.UserContract">Пользовательское соглашение</a></h3>
+    <h3>@Html.ActionLink("Пользовательское соглашение", "user-agreement", "OnlinePayments")</h3>
     <p>
         Документ регулирует отношения с пользователями JoinRpg — участниками мероприятий,
         и определяет правила регистрации на сервисе и работы с ним, в том числе
@@ -95,7 +95,7 @@
 </div>
 
 <div>
-    <h3><a href="@Documents.OrganizerContract">Агентский договор с юридическим лицом или индивидуальным предпринимателем</a></h3>
+    <h3>@Html.ActionLink("Агентский договор с юридическим лицом или индивидуальным предпринимателем", "organizer-contract", "OnlinePayments")</h3>
     <p>
         Этот договор необходимо заключить, если вы хотите собирать организационные взносы путем
         продажи билетов на свое мероприятие при помощи нашего сервиса от имени юридического лица
